Restructure scanners. Split functions into new files, use consistent format for class names
