+++ services/scanners/ssl/Dockerfile
@@ -17,4 +17,4 @@ RUN apt-get update && apt-get install -y --no-install-recommends \
 RUN pip3 install -r requirements.txt
 
 # Run the web service on container startup. Using uvicorn, in this case.
-CMD exec gunicorn ssl_scanner:app -w 4 --timeout 1000 -k uvicorn.workers.UvicornWorker
+CMD exec gunicorn server:app -w 4 --timeout 1000 -k uvicorn.workers.UvicornWorker
